// SDKs/bitbucket.org/2.0/typescript/src/sdk/models/shared/paginatedusers.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


// PaginatedUsers
/** 
 * A paginated list of users.
**/
export class PaginatedUsers extends SpeakeasyBase {
  @Metadata({ data: "json, name=next" })
  next?: string;

  @Metadata({ data: "json, name=page" })
  page?: number;

  @Metadata({ data: "json, name=pagelen" })
  pagelen?: number;

  @Metadata({ data: "json, name=previous" })
  previous?: string;

  @Metadata({ data: "json, name=size" })
  size?: number;

  @Metadata({ data: "json, name=values" })
  values?: Map<string, any>[];
}